refactor: extract mod peer_list
