+++ packages/torrent-repository/src/repository/skip_map_mutex_std.rs
@@ -9,7 +9,8 @@ use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
 use torrust_tracker_primitives::{peer, DurationSinceUnixEpoch, PersistentTorrents};
 
 use super::Repository;
-use crate::entry::{Entry, EntrySync, PeerList};
+use crate::entry::peer_list::PeerList;
+use crate::entry::{Entry, EntrySync};
 use crate::{EntryMutexStd, EntrySingle};
 
 #[derive(Default, Debug)]
